Fix numeric price sorting in OpenSearch and Redshift.
Remove alphanumeric sorting on cost columns and sort by numeric
cost values instead. Extract calculateCostNumeric for the sort accessor.
