+++ next/utils/colunnData/opensearch.tsx
@@ -1,6 +1,7 @@
 import { CostDuration, PricingUnit } from "@/types";
 import {
-    getPricingSorter,
+    calculateCost,
+    calculateCostNumeric,
     regex,
     makeCellWithRegexSorter,
     expr,
@@ -76,6 +77,49 @@ export function makePrettyNames<V>(
     ] as const;
 }
 
+function getPricingSorter(
+    selectedRegion: string,
+    pricingUnit: PricingUnit,
+    costDuration: CostDuration,
+    getter: (
+        pricing: OpenSearchPricing[string] | undefined,
+    ) => string | undefined,
+    currency: {
+        code: string;
+        usdRate: number;
+        cnyRate: number;
+    },
+) {
+    return {
+        sortingFn: "basic" as const,
+        sortUndefined: "last",
+        accessorFn: (row) => {
+            const g = getter(row.pricing?.[selectedRegion]);
+            if (isNaN(Number(g)) || !g) return undefined;
+            return calculateCostNumeric(
+                g,
+                row,
+                pricingUnit,
+                costDuration,
+                selectedRegion,
+                currency,
+            );
+        },
+        ...makeCellWithRegexSorter("pricing", (info) => {
+            const pricing = info.row.original.pricing;
+            const price = getter(pricing?.[selectedRegion]);
+            if (isNaN(Number(price)) || !price) return undefined;
+            return calculateCost(
+                price,
+                info.row.original,
+                pricingUnit,
+                costDuration,
+                selectedRegion,
+                currency,
+            );
+        }),
+    } satisfies Partial<ColumnDef<Instance>>;
+}
 
 export const columnsGen = (
     selectedRegion: string,
@@ -161,7 +205,7 @@ export const columnsGen = (
         accessorKey: "pricing",
         id: "cost-ondemand",
         header: "On Demand Cost",
-        ...getPricingSorter<Instance>(
+        ...getPricingSorter(
             selectedRegion,
             pricingUnit,
             costDuration,
@@ -173,7 +217,7 @@ export const columnsGen = (
         accessorKey: "pricing",
         id: "cost-reserved",
         header: "Reserved Cost",
-        ...getPricingSorter<Instance>(
+        ...getPricingSorter(
             selectedRegion,
             pricingUnit,
             costDuration,
